+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 **1.0.0.2020** *(May 11, 2018)*
 
 - Added extra vetting for security IDs.
+- Made sure that only login managers can see logins other than themselves, even when they have the token.
 
 **1.0.0.2019** *(May 10, 2018)*